requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e .
absl-py==1.2.0; python_version >= '3.6'
alembic==1.8.1; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0; python_version >= '3.6'
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0; python_version >= '3.5'
backcall==0.2.0
beautifulsoup4==4.11.1; python_version >= '3.6'
bleach==5.0.1; python_version >= '3.7'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3
cloudpickle==2.1.0; python_version >= '3.6'
colorama==0.4.5
commonmark==0.9.1
configparser==5.2.0
contextlib2==21.6.0; python_version >= '3.6'
cuda-python==11.7.1
cycler==0.11.0; python_version >= '3.6'
cython==0.29.32; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
databricks-cli==0.17.2
datetime==4.5
debugpy==1.6.3; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.3.5
docker==5.0.3; python_version >= '3.6'
entrypoints==0.4; python_version >= '3.6'
execnet==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.8.0; python_version >= '3.7'
filetype==1.1.0
flask==2.2.2; python_version >= '3.7'
flatbuffers==1.12
fonttools==4.36.0; python_version >= '3.7'
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.27; python_version >= '3.7'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==2.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
grpcio==1.47.0; python_version >= '3.6'
gunicorn==20.1.0; platform_system != 'Windows'
h5py==3.7.0
idna==3.3; python_version >= '3.5'
imbalanced-learn==0.9.1
importlib-metadata==4.12.0; python_version < '3.9'
importlib-resources==5.9.0; python_version < '3.9'
iniconfig==1.1.1
install==1.3.5
ipykernel==6.15.1
ipython-genutils==0.2.0
ipython==8.4.0; python_version >= '3.8'
ipywidgets==8.0.1
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.18.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.1.0; python_version >= '3.6'
jsonschema==4.13.0; python_version >= '3.7'
jupyter-client==7.3.4; python_version >= '3.7'
jupyter-console==6.4.4; python_version >= '3.7'
jupyter-core==4.11.1; python_version >= '3.7'
jupyter==1.0.0
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-widgets==3.0.1; python_version >= '3.7'
keras-preprocessing==1.1.2
keras-tqdm==2.0.1
keras==2.9.0
kiwisolver==1.4.4; python_version >= '3.7'
libclang==14.0.6
llvmlite==0.39.0; python_version >= '3.7'
lxml==4.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.2.1; python_version >= '3.7'
markdown==3.4.1; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
matplotlib==3.5.3
mistune==0.8.4
mlflow==1.28.0
mock==4.0.3; python_version >= '3.6'
nbclient==0.6.6; python_version >= '3.7'
nbconvert==6.5.3; python_version >= '3.7'
nbformat==5.4.0; python_version >= '3.7'
nest-asyncio==1.5.5; python_version >= '3.5'
notebook==6.4.12; python_version >= '3.7'
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.0; python_version >= '3.6'
opencv-python==4.6.0.66
opt-einsum==3.3.0; python_version >= '3.5'
packaging==21.3; python_version >= '3.6'
pandas==1.4.3
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
path.py==12.5.0; python_version >= '3.5'
path==16.4.0
pdf2image==1.16.0
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.2.0
pipenv==2022.8.19
pipfile==0.0.2
pkgutil-resolve-name==1.3.10; python_version < '3.9'
platformdirs==2.5.2; python_version >= '3.7'
pluggy==1.0.0; python_version >= '3.6'
prometheus-client==0.14.1; python_version >= '3.6'
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.30; python_full_version >= '3.6.2'
protobuf==3.19.4; python_version >= '3.5'
psutil==5.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
pydot==1.4.2
pygments==2.13.0; python_version >= '3.6'
pyjwt==2.4.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
pytest-shutil==1.7.0
pytest==7.1.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2022.2.1
pyyaml==6.0
pyzmq==23.2.1; python_version >= '3.6'
qtconsole==5.3.1; python_version >= '3.7'
qtpy==2.2.0; python_version >= '3.7'
querystring-parser==1.2.4
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1
rich==12.5.1
rsa==4.9; python_version >= '3.6'
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
send2trash==1.8.0
setuptools==65.1.0
shellingham==1.5.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sklearn==0.0
smmap==5.0.0; python_version >= '3.6'
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlalchemy==1.4.40; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.2; python_version >= '3.5'
stack-data==0.4.0
tabulate==0.8.10
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.1; python_version >= '3.6'
tensorflow-addons==0.17.1
tensorflow-estimator==2.9.0; python_version >= '3.7'
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0; python_version < '3.11' and python_version >= '3.7'
tensorflow==2.9.1
termcolor==1.1.0
terminado==0.15.0; python_version >= '3.7'
tf-explain==0.3.1
threadpoolctl==3.1.0; python_version >= '3.6'
tifffile==2022.8.12
tinycss2==1.1.1; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tomli==2.0.1; python_version >= '3.7'
tornado==6.2; python_version >= '3.7'
tqdm==4.64.0
traitlets==5.3.0; python_version >= '3.7'
typeguard==2.13.3; python_full_version >= '3.5.3'
typer[all]==0.6.1
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
virtualenv-clone==0.5.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.16.3; python_version >= '3.6'
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3; python_version >= '3.7'
werkzeug==2.2.2; python_version >= '3.7'
wheel==0.37.1
widgetsnbextension==4.0.1; python_version >= '3.7'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.8.1; python_version >= '3.7'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
